requirements.txt
langchain-openai
langchain
python-dotenv
ipykernel
langchain_community
langchain_core
streamlit
langchain-huggingface
transformers
langchain-groq
pypdf
langchain-ollama
faiss-cpu
langchain_chroma
chromadb
arxiv
wikipedia
faiss-cpu